refactor: group metrics and tracing under telemetry/
Move the two OpenTelemetry exporters (metrics/ and tracing/) under a single telemetry/ umbrella to express that they are the export layer. observers/ stays top-level as the in-process pipeline mechanism. Package names are unchanged, so call sites (metrics.RecordTTFB, tracing.Tracer, ...) are identical — only import paths move. Configs are left independent (they share only four trivial, stable OTLP fields; tracing additionally has SampleRatio).
