+++ service/llm/llm.go
@@ -24,9 +24,9 @@ import (
 	"time"
 
 	"github.com/gojargo/jargo/frames"
-	"github.com/gojargo/jargo/metrics"
 	"github.com/gojargo/jargo/processor"
-	"github.com/gojargo/jargo/tracing"
+	"github.com/gojargo/jargo/telemetry/metrics"
+	"github.com/gojargo/jargo/telemetry/tracing"
 	"go.opentelemetry.io/otel/attribute"
 	"go.opentelemetry.io/otel/trace"
 )
